# backend/contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabelNodeIn": {
        "properties": {
          "category_id": {
            "title": "Category Id",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "default_prompt_version": {
            "default": "v1",
            "title": "Default Prompt Version",
            "type": "string"
          },
          "is_leaf": {
            "default": false,
            "title": "Is Leaf",
            "type": "boolean"
          },
          "level": {
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          },
          "llm_enabled": {
            "default": true,
            "title": "Llm Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "required": [
          "category_id",
          "level",
          "name",
          "code"
        ],
        "title": "LabelNodeIn",
        "type": "object"
      },
...
      "LabelTaxonomyIn": {
        "properties": {
          "businessScope": {
            "items": {
              "type": "string"
            },
            "title": "Businessscope",
            "type": "array"
          },
          "categoryScope": {
            "items": {
              "type": "string"
            },
            "title": "Categoryscope",
            "type": "array"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "default": "",
            "title": "Owner",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "code"
        ],
        "title": "LabelTaxonomyIn",
        "type": "object"
      },
...
      "LabelTaxonomyOut": {
        "properties": {
          "businessScope": {
            "items": {
              "type": "string"
            },
            "title": "Businessscope",
            "type": "array"
          },
          "categoryScope": {
            "items": {
              "type": "string"
            },
            "title": "Categoryscope",
            "type": "array"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "currentVersionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentversionid"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nodeCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": 0,
            "title": "Nodecount"
          },
          "owner": {
            "default": "",
            "title": "Owner",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "title": "Status",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedby"
          }
        },
        "required": [
          "name",
          "code",
          "id",
          "createdAt",
          "updatedAt"
        ],
        "title": "LabelTaxonomyOut",
        "type": "object"
      },
      "LabelTaxonomyVersionOut": {
        "properties": {
          "changeLog": {
            "default": "",
            "title": "Changelog",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "nodeCount": {
            "default": 0,
            "title": "Nodecount",
            "type": "integer"
          },
          "publishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "taxonomyId": {
            "title": "Taxonomyid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedby"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "taxonomyId",
          "version",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "LabelTaxonomyVersionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "VOC Labeling Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/label-taxonomies": {
      "get": {
        "operationId": "list_taxonomies_api_v1_label_taxonomies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabelTaxonomyOut"
                  },
                  "title": "Response List Taxonomies Api V1 Label Taxonomies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Taxonomies",
        "tags": [
          "label-taxonomies"
        ]
      },
      "post": {
        "operationId": "create_taxonomy_api_v1_label_taxonomies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelTaxonomyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelTaxonomyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Taxonomy",
        "tags": [
          "label-taxonomies"
        ]
      }
    },
    "/api/v1/label-taxonomies/{taxonomy_id}": {
      "get": {
        "operationId": "get_taxonomy_api_v1_label_taxonomies__taxonomy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "taxonomy_id",
            "required": true,
            "schema": {
              "title": "Taxonomy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelTaxonomyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Taxonomy",
        "tags": [
          "label-taxonomies"
        ]
      },
      "put": {
        "operationId": "update_taxonomy_api_v1_label_taxonomies__taxonomy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "taxonomy_id",
            "required": true,
            "schema": {
              "title": "Taxonomy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelTaxonomyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelTaxonomyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Taxonomy",
        "tags": [
          "label-taxonomies"
        ]
      }
    },
    "/api/v1/label-taxonomies/{taxonomy_id}/versions": {
      "get": {
        "operationId": "list_taxonomy_versions_api_v1_label_taxonomies__taxonomy_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "taxonomy_id",
            "required": true,
            "schema": {
              "title": "Taxonomy Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabelTaxonomyVersionOut"
                  },
                  "title": "Response List Taxonomy Versions Api V1 Label Taxonomies  Taxonomy Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Taxonomy Versions",
        "tags": [
          "label-taxonomies"
        ]
      }
    },
    "/api/v1/label-taxonomies/{taxonomy_id}/versions/{version_id}": {
      "get": {
        "operationId": "get_taxonomy_version_api_v1_label_taxonomies__taxonomy_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "taxonomy_id",
            "required": true,
            "schema": {
              "title": "Taxonomy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelTaxonomyVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Taxonomy Version",
        "tags": [
          "label-taxonomies"
        ]
      }
    },
    "/api/v1/label-taxonomies/{taxonomy_id}/versions/{version_id}/tree": {
      "get": {
        "operationId": "get_taxonomy_tree_api_v1_label_taxonomies__taxonomy_id__versions__version_id__tree_get",
        "parameters": [
          {
            "in": "path",
            "name": "taxonomy_id",
            "required": true,
            "schema": {
              "title": "Taxonomy Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabelTaxonomyNodeOut"
                  },
                  "title": "Response Get Taxonomy Tree Api V1 Label Taxonomies  Taxonomy Id  Versions  Version Id  Tree Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Taxonomy Tree",
        "tags": [
          "label-taxonomies"
        ]
      }
    },
    "/api/v1/labels": {
      "post": {
        "operationId": "create_label_api_v1_labels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelNodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelNodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Label",
        "tags": [
          "labels"
        ]
      }
    },
...
    "/api/v1/labels/{label_id}": {
      "delete": {
        "operationId": "delete_label_api_v1_labels__label_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "title": "Label Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Label Api V1 Labels  Label Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Label",
        "tags": [
          "labels"
        ]
      },
      "get": {
        "operationId": "get_label_api_v1_labels__label_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "title": "Label Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelNodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Label",
        "tags": [
          "labels"
        ]
      },
      "put": {
        "operationId": "update_label_api_v1_labels__label_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "label_id",
            "required": true,
            "schema": {
              "title": "Label Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelNodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelNodeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Label",
        "tags": [
          "labels"
        ]
      }
    },
...
  }
}